Refactor add-to-cart button into reusable component
Extracted the add-to-cart button logic into a new ButtonAddToCart component and replaced inline implementations in ProductCard and ProductIdView. Updated product links to use slug instead of id for consistency.
